PATCH fixes, added functionality
fixed PATCH issue of deleting values of attributes that were omitted from a request. Added functionality to update customfields, filters, and instances, and also add instances. Also implemented a rudementary DELETE instances function, whereby setting starttime and endtime to null soft deletes the instance. This is similar to how JSONapi specifies to delete a relationship, although with regards to relationships, this API isn't totally in line with JSONapi spec.