add channels only field to feed
Also changes the valid visibility fields. Right now, 'Channels' is the only visibility field being used so this won't be a breaking change. For localist 'Unlisted' and 'Restricted' are the only valid visibility values, but I kept 'Channels' for backward compatibility. If the visibility is 'Channels', the 'Channels Only' CSV field will be set to 1 (true).